Refactor baseline
