+++ core/testcontainers/core/utils.py
@@ -1,17 +1,12 @@
-import base64
-import json
 import logging
 import os
 import platform
 import subprocess
 import sys
-from collections import namedtuple
 
 LINUX = "linux"
 MAC = "mac"
 WIN = "win"
-
-DockerAuthInfo = namedtuple("DockerAuthInfo", ["registry", "username", "password"])
 
 
 def setup_logger(name: str) -> logging.Logger:
@@ -82,29 +77,3 @@ def raise_for_deprecated_parameter(kwargs: dict, name: str, replacement: str) ->
     if kwargs.pop(name, None):
         raise ValueError(f"Use `{replacement}` instead of `{name}`")
     return kwargs
-
-
-def parse_docker_auth_config(auth_config: str) -> list[DockerAuthInfo]:
-    """
-    Parse the docker auth config from a string.
-
-    Example:
-    {
-        "auths": {
-            "https://index.docker.io/v1/": {
-                "auth": "dXNlcm5hbWU6cGFzc3dvcmQ="
-            }
-        }
-    }
-    """
-    auth_info: list[DockerAuthInfo] = []
-    try:
-        auth_config_dict: dict = json.loads(auth_config).get("auths")
-        for registry, auth in auth_config_dict.items():
-            auth_str = auth.get("auth")
-            auth_str = base64.b64decode(auth_str).decode("utf-8")
-            username, password = auth_str.split(":")
-            auth_info.append(DockerAuthInfo(registry, username, password))
-        return auth_info
-    except (json.JSONDecodeError, KeyError, ValueError) as exp:
-        raise ValueError("Could not parse docker auth config") from exp
